Modify Makefile to work with recent make
The `proxies:..` block works on 3.x makes, but
not with make 4.1. This modifies it to work on both
older and more recent versions of make.

See https://stackoverflow.com/questions/34562221/recipe-commences-before-first-target-stop/43393843
for some details on the specific error seen on make 4.1.

TBH I don't know what the exact syntax problem was - but I do know
that I can run `make` and `make container` after applying this
diff.