notifications: Add new $kubemonkeyid placeholder
Sometimes it's possible to run kube-monkey on multiple clusters
but all of them are sending notification so the same endpoint. In order
cases, we may run difference instances of kube-monkey on the same
cluster. It's useful to be able to identify the schedules and
notifications attacks and be able to correlate them with a particular
kube-monkey instance. As such, a new $id placeholder is added which
takes its value from the KUBE_MONKEY_ID env variable.